BDIA-final Project/requirements.txt
fastapi>=0.104.1
starlette>=0.27.0
uvicorn==0.24.0
python-dotenv==1.0.0
pinecone-client==2.2.4
langchain
openai>=1.40.0,<2.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
python-multipart==0.0.6
fastapi-limiter==0.1.5
redis==4.5.4
aiohttp>=3.8.6,<4.0.0
langsmith
langchain-core
langchain-openai
langgraph>=0.0.10
google-cloud-bigquery
google-api-python-client
requests